flexbox and some css styling added to replace semantic-ui-react.  Results are inconsistent at best.  Unsure if it's the dev tools or the code.  Currently seems to be working on mobile sized for iPhone5 and desktop.  Resutls pages not yet styled.
