+++ src/index.css
@@ -1,11 +1,39 @@
 body {
   margin: 0;
-  padding: 0;
   font-family: sans-serif;
   list-style: none;
+  box-sizing: border-box;
 }
-.result-image {
-  width: 200px;
-  height: 200px;
+.container {
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  padding: 4px;
+}
+.flex-item {
+  align-self: center;
+  flex: 0 1 auto;
+}
+.center {
+  text-align: center;
+}
+.spacer {
+  margin-top: 2.85714286em;
+}
+.block {
+  display: block;
+}
+.rounded {
+  border: 1px solid #22426;
+  border-radius: .28571429rem;
+  min-height: 2.85714286em;
+  box-shadow: 0 1px 2px 0 #22426;
+}
+.button {
+  min-width: 174px;
+  margin-top: 4px;
+}
+@media all and (min-width: 720px) {
+  max-width: 100%;
 }
