+++ src/modules/home/StarRater.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import '../../index.css';
 
 function StarRater(props) {
     const stars = [];
