+++ src/modules/home/vis-result-list.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ResultList from './result-list';
 import {connect} from 'react-redux';
 import * as actions from '../../redux/actions';
+import '../../index.css';
 
 class ReturnedResultList extends React.Component {
   render() {
